Tab one: first attempt `add and complete task` functionality
Write the task inside of `TextInput` and add task by pressing circular
`+` button. To complete the task, press the any where in the task box
(not the blue box in the left, that box still have no function). And
also the title changed if there's at least one task or not.