Task components: first attempt delete button
Almost the same as `complete` function but didn't store the item.
There's still an issue when pressing the same character. For example:
when the user type `44`, then the function won't delete the task because
the function think that there's two children with the same key (4). Not
sure how to fix this at the moment.