Task components: make `square button` as `complete button`
Previously, the square button (or to be precise `TouchableOpacity`) do
nothing. Complete task done by pressing the main element of task list,
and now you need to press the square button to complete the task. At the
moment, it still the remove the task but stored in `completed` array. Need
to figure out how to cross the completed mark instead of remove it from
the list.