Add new `delete task` functionality
The previous `delete task` function delete task based on index. Now that
the unique `key` for task list already implemented, use the unique `key`
to delete (or to be precise filter) the task item. There's an issue with
this method, if the virtual keyboard still on screen, user can't delete
the task directly. When user press the `delete button`, it will close
the virtual keyboard first. After the virtual keyboard closed, user
can delete the task item.